Add a workaround for a scoring issue in pump mode
There is an inconsistency in the game engine in pump game mode that when
PercentScoreWeightCheckpointHit and PercentScoreWeightHeld are both
non-zero even perfect play can lead to a score <100%. The workaround is
to set PercentScoreWeightHeld to 0 in pump mode. The workaround is also
applied in SM5's fallback theme.

Issue reported by @SheepyChris in
https://github.com/Simply-Love/Simply-Love-SM5/pull/222